requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-data-tables>=12.4.0
pandas>=1.5.0
openpyxl>=3.0.0
pandas
pydantic
openai